database-setup-test/requirements.txt
cassandra-driver
cassio
transformers
sentence-transformers